Add initial Temperature class

+++ src/WDataPoint.java
@@ -12,8 +12,7 @@
  */
 public interface WDataPoint {
     /** Add a data point from the sensor to the cumulative total tracked by the object. */
-    void addDataPoint(float point); // should this be float? other data type?
+    void addDataPoint(double point); // should this be double? other data type?
     /** Convert the fields of the object to an array of bytes. */
     byte[] toByteArray(); // what about deserialization?
-    // how should historical data be returned by these objects?
 }
